pyproject.toml
[project]
name = "brooks-sla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "emgardner", email = "gardner.ethan10@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "hart-protocol>=2023.6.0",
    "pydantic>=2.12.5",
    "pyserial-asyncio>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
    "ty>=0.0.1a35",
]
